stacks/amilifecycle/resources/lambda/ami_lifecycle/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
requests==2.27.1
urllib3==1.26.7
Jinja2==3.0.3
MarkupSafe==2.0.1
PyYAML==5.4.1
